code complete: change float to int

+++ batch_1/prog_4_print_product.py
@@ -1,8 +1,8 @@
 # Prog04: Create a program that ask user to input 2 numbers. Print the product of the two numbers.
 
 # ask user two numbers 
-num1 = float(input ("Enter first number: "))
-num2 = float(input ("Enter second number: "))
+num1 = int(input ("Enter first number: "))
+num2 = int(input ("Enter second number: "))
 result = num1 * num2
 
 # print product
